add(Security): tests that provide more checks on security issues
like trying to access plants that one does not own

+++ tests/Controller/LoginControllerTest.php
@@ -20,7 +20,7 @@ class LoginControllerTest extends WebTestCase
 
         // Remove any existing users from the test database
         foreach ($userRepository->findAll() as $user) {
-            if($user->getUsername() !== 'Testuser'){
+            if($user->getUsername() !== 'Testuser' && $user->getUsername() !== 'TestuserNoRef') {
                 $em->remove($user);
             }
         }
